Prototype optional key for nearest neighbour
